application/cbf_crawler/requirements.txt
bs4==0.0.1
requests==2.26.0
psycopg2-binary==2.9.1
pyyaml==5.4.1